+++ src/main/java/de/samply/result/container/Containers.java
@@ -1,20 +1,13 @@
 package de.samply.result.container;
 
 import de.samply.result.container.template.ContainerTemplate;
-import de.samply.result.container.template.ContainersTemplate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class Containers {
-
-  private ContainersTemplate containersTemplate;
   private Map<ContainerTemplate, Map<String, Container>> templateIdContainerMap = new HashMap<>();
-
-  public Containers(ContainersTemplate containersTemplate) {
-    this.containersTemplate = containersTemplate;
-  }
 
   public void addContainer(Container container) {
     Map<String, Container> idContainerMap = templateIdContainerMap.get(
@@ -34,10 +27,6 @@ public class Containers {
   public List<Container> getContainers(ContainerTemplate containerTemplate) {
     Map<String, Container> idContainerMap = templateIdContainerMap.get(containerTemplate);
     return (idContainerMap != null) ? new ArrayList<>(idContainerMap.values()) : new ArrayList<>();
-  }
-
-  public ContainersTemplate getContainersTemplate() {
-    return containersTemplate;
   }
 
 }