+++ src/test/java/de/samply/container/template/ResultTemplateManagerTest.java
@@ -1,19 +1,19 @@
 package de.samply.container.template;
 
-import de.samply.result.container.template.ResultTemplatesManager;
+import de.samply.result.container.template.ResultTemplateManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 @Disabled
-class ResultTemplatesManagerTest {
+class ResultTemplateManagerTest {
 
-  private ResultTemplatesManager templates;
+  private ResultTemplateManager templates;
   private String templatesDirectory = "./templates";
 
   @BeforeEach
   void setUp() {
-    this.templates = new ResultTemplatesManager(templatesDirectory);
+    this.templates = new ResultTemplateManager(templatesDirectory);
     //TODO
   }